Add files via upload
An expense tracker, storing the date to a Mysql database on kapsi.fi server.  Outcommented some rows, where I tested it with XAMPP localhost, where the Mysql database looks a little bit different. This is quite messy php code, and I used it for self-teaching, There is a header and a footer, inside of a template directory, this is not needed for functionality. When downloading the mysql database, I export only kuukausi_1, the others are analogous kuukausi_2,  ... .kuukausi_12 (kuukausi=month), and the vuosi (year)